pyproject.toml
[project]
name = "autom-weblogistica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12",
  "imap-tools>=1.10.0",
  "imapclient>=3.0.1",
  "playwright>=1.52.0",
  "python-dotenv>=1.1.0",
  "uvicorn>=0.34.2",
]


[tool.basedpyright]
typeCheckingMode = "standard"
